moved js logic to own file
